feat: add projects

+++ deployment_back/src/project/project.service.ts
@@ -1,90 +1,98 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateProjectDto } from './dto/create-project.dto';
-import { UpdateProjectDto } from './dto/update-project.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Project } from '../project/project.entity';
 import { Repository } from 'typeorm';
 import { GetProjectDto } from './dto/get-project.dto';
 import { normalizeProjectName } from 'src/utils';
-import { SshProvider } from 'src/ssh/ssh.provider';
 import { FileEncryptorProvider } from '../file-encryptor/file-encryptor.provider';
-import { MiniBackService } from 'src/mini-back/mini-back.service';
+import fs from 'fs/promises';
+import { UpdateProjectDto } from './dto/update-project.dto';
+import { HttpService } from '@nestjs/axios';
 
 @Injectable()
 export class ProjectService {
   constructor(
     @InjectRepository(Project) private projectRepository: Repository<Project>,
-    private sshProvider: SshProvider,
     private fileEncryptorProvider: FileEncryptorProvider,
-    private miniBackService: MiniBackService,
+    private readonly httpService: HttpService,
   ) {}
 
   async create(
     createProjectDto: CreateProjectDto,
     envFilePath: string,
-    sshServerPrivateKeyPath: string,
     sshGitPrivateKeyProjectPath: string,
-    sshGitPublicKeyProjectPath: string,
   ) {
+    const candidate = await this.findAll({
+      miniBackId: createProjectDto.miniBackId,
+      name: createProjectDto.name,
+    });
+
+    if (candidate.length > 0) {
+      throw new ConflictException('Try to make name of a project unique');
+    }
+
     createProjectDto.name = normalizeProjectName(createProjectDto.name);
 
     await this.fileEncryptorProvider.encryptFilesOnPlace([
-      sshServerPrivateKeyPath,
       sshGitPrivateKeyProjectPath,
-      sshGitPublicKeyProjectPath,
       envFilePath,
     ]);
 
-    sshServerPrivateKeyPath = sshServerPrivateKeyPath + '.enc';
     sshGitPrivateKeyProjectPath = sshGitPrivateKeyProjectPath + '.enc';
-    sshGitPublicKeyProjectPath = sshGitPublicKeyProjectPath + '.enc';
     envFilePath = envFilePath + '.enc';
 
     const newProject = this.projectRepository.create({
       ...createProjectDto,
       envFilePath,
-      sshServerPrivateKeyPath,
       sshGitPrivateKeyProjectPath,
-      sshGitPublicKeyProjectPath,
     });
 
     return this.projectRepository.save(newProject);
   }
 
-  async deploy(projectId: string) {
-    const project = await this.findOne(projectId);
-
+  async deployProject(id: string) {
+    const project = await this.findOne(id);
     if (project === null) {
-      throw new NotFoundException("That project was't found");
+      throw new NotFoundException('Such project was not found');
     }
 
-    const status = await this.miniBackService.pingMiniBack(project.serverUrl);
-
-    if (!status) {
-      // const { serverUrl, sshServerPrivateKeyPath, gitProjectLink } = project;
-      // await this.miniBackService.placeMiniBake({
-      //   serverUrl,
-      //   sshServerPrivateKeyPath,
-      //   gitProjectLink,
-      // });
-    }
-
+    // const body = {};
+    // axios.post();
+    this.update(project.id, { isDeploy: true });
     return 'success';
+  }
+
+  async update(id: string, parameters: UpdateProjectDto) {
+    const project = await this.findOne(id);
+    if (project === null) {
+      throw new NotFoundException('Such a project was not found');
+    }
+
+    this.projectRepository.update({ id }, parameters);
   }
 
   async findAll(getProjectDto: GetProjectDto) {
     return this.projectRepository.findBy(getProjectDto);
   }
 
-  findOne(id: string) {
+  async findOne(id: string) {
     return this.projectRepository.findOneBy({ id });
   }
 
-  update(id: string, updateProjectDto: UpdateProjectDto) {
-    throw new Error('not implement yet');
-  }
+  async remove(id: string) {
+    const project = await this.findOne(id);
+    if (project === null) {
+      throw new NotFoundException('Not found such a project');
+    }
 
-  remove(id: string) {
+    await fs.unlink(project.envFilePath);
+    await fs.unlink(project.sshGitPrivateKeyProjectPath);
+
     return this.projectRepository.delete({ id });
   }
 }
